Minimal time required to cover earth, 1st take
Added calc to find the minimal time that would be required to try
and cover the earth. It tries to find how many orbits at a minimum
is required to cover the full of earth's equator.

It doesnt worry about how much of ground can be functionaly/practically
serviced by the satellite, in each orbit. Nor does it worry about how
the orbits may overlap or not and other nitty gritties.

NOTE: The overly crude eclipse logic added yesterday just gives the
a kind of maximal outer limit wrt eclipse time, the real time should
be much smaller.

NOTE: As can be noted above, for one of the parameters I have tried
to get the minimalest time required, while for the other I have
tried to get the maximal time required. And these are using crude/
simpel minded geometric calcs.

NOTE: Had forgotten to add the jupyter notebook yesterday, have
fixed it today.